Adjust damage storage to long.

+++ morg-api/src/main/java/com/atlas/morg/rest/attribute/DamageEntry.java
@@ -1,4 +1,4 @@
 package com.atlas.morg.rest.attribute;
 
-public record DamageEntry(int characterId, int damage) {
+public record DamageEntry(int characterId, long damage) {
 }
