+++ morg-api/src/main/java/com/atlas/morg/rest/event/DamageEntry.java
@@ -1,4 +1,4 @@
 package com.atlas.morg.rest.event;
 
-public record DamageEntry(int character, int damage) {
+public record DamageEntry(int character, long damage) {
 }
