+++ morg-model/src/main/java/com/atlas/morg/model/DamageEntry.java
@@ -1,4 +1,4 @@
 package com.atlas.morg.model;
 
-public record DamageEntry(int characterId, int damage) {
+public record DamageEntry(int characterId, long damage) {
 }
