+++ morg-model/src/main/java/com/atlas/morg/model/DamageSummary.java
@@ -1,4 +1,4 @@
 package com.atlas.morg.model;
 
-public record DamageSummary(int characterId, int uniqueId, int visibleDamage, int actualDamage, boolean killed) {
+public record DamageSummary(int characterId, int uniqueId, long visibleDamage, long actualDamage, boolean killed) {
 }
